Check if folder models exist
Hi there,

After trying your plugin i discovered a bug. I have some plugins that don't use models, In cases like this we get an error because we check for files in an non existing folder.

By checking if the folder exists we can skip the plugin instead of getting an error. Seams better that way ;)